backend/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14,<3.15
pandas>=2.0,<3
openpyxl>=3.1,<4
